Southern Elephant seal tag data

Light data from the foraging trips of the Southern elephant seal.

Format

ElephantSeal1 A data frame with 3 columns. The columns represent

time 	 times of measurement

light 	 (log) light values

segment 	 integer indicating sequence of twilight periods

ElephantSeal2 This tag is similar to ElephantSeal1 but also has columns

depth 	 depth in the water column in metres (positive)

temp 	 temperature in the water column in degrees celcius

ElephantSeal1calib and ElephantSealcalib2 A data frame with 2 columns. The columns represent

zenith 	 zenith values

light 	 (log) light values

Details

Light intensity measurements over time from an archival tag on a Southern elephant seal (Mirounga leonina). The seal was tagged at the isthmus on Macquarie Island (158.950E, 54.5S), with data from a time-depth-recorder (Mk9 TDR; Wildlife Computers, Seattle, WA, USA). These tags provides regular time series of measurements of depth, water temperature, and ambient light level. The original data for ElephantSeal1 were processed to remove values at depths greater than 15m and to classify periods of twilight. The data for ElephantSeal2 have also been processed for twilight periods. The seals makes one single foraging trip, returning to the isthmus where they were tagged. Data recorded while the seal is at the isthmus are used for calibration (ElephantSeal1calib).

These data supplied courtesy of Mark Hindell, Institute of Marine and Antarctic Studies, University of Tasmania. ElephantSeal1 is B362_99 and ElephantSeal2 is C699_02