Short-tailed Shearwater Track

Twilight times from a flight of a Short-tailed Shearwater.

Format

A data frame with 3 columns and 72 rows. The columns represent

twilight 	 times of twilight.

rise 	 logical indicating which twilights are sunrise.

Details

Twilight times derived from light intensity measurements from an archival tag on a Short-tailed Shearwater (Puffinus tenuirostris). The bird was tagged at its burrow on Wedge Island, Tasmania Australia (147.670E 43.133S). The bird makes two foraging trips, returning to its burrow after each trip.

When the bird is in its burrow, no light is recorded by the tag and the twilights reported in the data set are the calculated twilights Wedge Island when solar zenith is 95 degrees.

These data supplied courtesy of Delphi Ward and Mark Hindell, Institute of Marine and Antarctic Studies, University of Tasmania.