Geolocation Estimation by the Threshold Method

Estimate location from two consecutive twilights by a threshold method

coord(tFirst, tSecond, type, degElevation = -6)

Arguments

tFirst
factor or character vector representing the times of the first twilight in the format "Y-m-d H:M:S".

tSecond
factor or character vector representing the times of the second twilight in the format "Y-m-d H:M:S".

type
vector with elements 1 or 2, defining the elements of tFirst as sunrise or sunset respectively.

degElevation
sun elevation angle (90-zenith).

Value

location estimates stored as a two column (lon,lat) matrix.

Details

This package and the GeoLight package provide some common functionality, but are based on different astronomical approximations. This function provides a drop-in replacement for the coord function from the GeoLight package to allow easy comparison of the two systems.