Metropolis Samplers

Metropolis samplers for Stella or Estelle.

estelleMetropolis(model, proposal.x, proposal.z, x0 = NULL, z0 = NULL,
  iters = 1000L, thin = 10L, chains = 1L, verbose = interactive())

stellaMetropolis(model, proposal.x, x0 = NULL, iters = 1000L, thin = 10L,
  chains = 1L, verbose = interactive())

Arguments

model
a model structure as generated by thresholdModel.

proposal.x
function for drawing proposals for x.

proposal.z
function for drawing proposals for z.

x0
Starting values for twilight locations x.

z0
Starting values for intermediate locations z.

iters
number of samples to draw.

thin
rate at which to thin samples.

chains
number of chains to sample.

verbose
report progress at prompt?

Value

If there are r samples drawn for each of q chains of p parameters at n locations, Stella will return a list containing

model

the model structure

x

a list of n x p x r arrays of twilight locations from the q chains

While in addition Estelle will return
z

a list of (n-1) x p x r arrays of intermediate locations from the q chains

.
Details

These functions draw samples form posterior for the Stella or Estelle model by the Metropolis algorithm.

See also

thresholdModel