Generalized Array Permutation

Permute and collapse the dimensions of an array

gperm(x, perm)

Arguments

x
an array.

perm
the subscript permutation list.

Value

The redimensioned array.

Details

Like aperm this function permutes the dimensions of an array, but can also collapse adjacent dimensions. The permutation of dimensions is specified by a list of vectors. The index of each dimension must appear exactly once, but dimensions that occur together in a vector are reduced to a single dimension.

Examples

x <- array(1:120,2:5)
dim(gperm(x,list(4,3,2,1)))

#> [1] 5 4 3 2

dim(gperm(x,list(c(4,3),2,1)))

#> [1] 20  3  2

dim(gperm(x,list(1,c(3,2),4)))

#> [1]  2 12  5