Summaries of Location Samples

Summarize a set of location samples

locationSummary(s, time = NULL, discard = 0, alpha = 0.95,
  collapse = TRUE, chains = NULL)

locationMean(s, discard = 0, collapse = TRUE, chains = NULL)

Arguments

s
a single chain or a list of parallel chains generated by estelleMetropolis or stellaMetropolis.

time
the times corresponding to the x locations.

discard
number of initial samples to discard.

alpha
coverage of the credible intervals calculated by locationSummary.

collapse
whether to collapse parallel chains to a single chain

chains
the set of chains to retain, or NULL.

Value

locationSummary

returns a dataframe or a list of dataframes of summary quantities for each location.

locationMean

returns an array or a list of arrays of the means of the samples for each location.

Details

These functions compute various summaries of a sample or list of samples generated by estelleMetropolis or stellaMetropolis.

These functions accept either a sample from a single mcmc run, or a list of samples from parallel mcmc runs. When collapse is true, multiple samples are merged and single result is returned, otherwise a result is returned for each sample.