Twilight Error Models

Log density of twilight errors

makeTwilightModel(twilight.model = c("Gamma", "LogNormal", "Normal",
  "ModifiedGamma", "ModifiedLogNormal"), alpha)

Arguments

twilight.model
the model for the errors in twilight times.

alpha
parameters of the twilight model.

Value

a function to evaluate the log density of the twilight residuals in a threshold model.

Details

Construct a function to evalute the log density of the twilight errors in a threshold model.

One of several models models may be selected for the errors in twilight times. The errors in twilight time are defined as the difference in the observed and true times of twilight, with sign selected so that a positive error always corresponds to a sunrise observed after the true time of sunrise, and sunset observed before the true time of sunset. That is, a positive error corresponds to the observed light level being lower than expected.

The properties of the twilight model are determined by alpha, which must be either a vector of parameters that are to be applied to each twilight, or a matrix of parameters with one row for each twilight.

The twilight.model argument selects the distribution of the twilight errors

'Normal'

Normally distributed with mean alpha[,1] and standard deviation alpha[,2],

'LogNormal'

Log Normally distributed so the log errors have mean alpha[,1] and standard deviation alpha[,2], or

'Gamma'

Gamma distributed with shape alpha[,1] and rate alpha[2].

The 'LogNormal' and 'Gamma' models forbid negative errors, that is, the observed light cannot be brighter than expected. There are modified variants of these models for which negative errors are extremely unlikely, but not forbidden, and can be used to generate suitable initialization locations for their unmodified counterparts.