Atmospheric Refraction

Adjust the solar zenith angle for atmospheric refraction.

refracted(zenith)

unrefracted(zenith)

Arguments

zenith
zenith angle (degrees) to adjust.

Value

vector of zenith angles (degrees) adjusted for atmospheric refraction.

Details

Given a vector of solar zeniths computed by zenith, refracted calculates the solar zeniths adjusted for the effect of atmospheric refraction. unrefracted is the inverse of refracted. Given a (single) solar zenith adjusted for the effect of atmospheric refraction, unrefracted calculates the solar zenith as computed by zenith.

Examples

## Refraction causes the sun to appears higher on the horizon
refracted(85:92)

#> [1] 84.84038 85.81110 86.77132 87.71632 88.63663 89.51806 90.66944 91.83477

## unrefracted gives unadjusted zenith
unrefracted(refracted(90))

#> $root
#> [1] 90
#> 
#> $f.root
#> [1] -8.318548e-08
#> 
#> $iter
#> [1] 5
#> 
#> $init.it
#> [1] NA
#> 
#> $estim.prec
#> [1] 6.103516e-05
#>