Satellite Model Structures

Satellite Model Structure for Stella and Estelle

satelliteModel(time, X, location.model = c("Normal", "T"), sd, df = NULL,
  beta, logp.x = function(x) rep.int(0L, nrow(x)), logp.z = function(z)
  rep.int(0L, nrow(z)), x0, z0 = NULL, fixedx = FALSE, dt = NULL)

satelliteModel0(time, X, location.model = c("Normal", "T"), sd, df = NULL,
  logp.x = function(x) rep.int(0L, nrow(x)), logp.z = function(z)
  rep.int(0L, nrow(z)), fixedx = FALSE)

Arguments

time
the times of the satellite determined locations.

X
the satellite determined locations.

location.model
the model for the errors in satellite locations.

sd
a vector or two column matrix of dispersions for the location model.

df
a vector or two column matrix of degrees of freedom for the t location model.

beta
parameters of the behavioural model.

logp.x
function to evaluate any additional contribution to the log posterior from the satellite estimated locations.

logp.z
function to evaluate any additional contribution to the log posterior from the intermediate locations.

x0
suggested starting points for the satellite locations.

z0
suggested starting points for intermediate locations.

fixedx
logical vector indicating which satellite locations to hold fixed.

dt
time intervals for speed calculation in hours.

Value

a list with components

logpx

function to evaluate the contributions to the log posterior from the twilight model

logpz

function to evaluate the contributions to the log posterior from the prior for the z locations

estelle.logpb

function to evaluate contribution to the log posterior from the behavioural model for estelle.

stella.logpb

function to evaluate contribution to the log posterior from the behavioural model for stella.

fixedx

a logical vector indicating which locations should remain fixed.

x0

an array of initial twilight locations.

z0

an array of initial intermediate locations.

time

the times of the satellite determined locations.

X

the satellite determined locations.

Details

Stella and Estelle require a model structure that describes the model being fitted. This function generates a basic model structure for satellite data that should provide a suitable starting point for most analyses.

The satelliteModel function constructs a model structure assuming that assumes the locations X determined by the satellite are independently distributed about the corresponding true locations. The location.model parameter selects whether (X-x)/sd is

'Normal'

Normally distributed with zero mean and unit variance, or

'T'

t distributed with degrees of freedom df.

Both Estelle and Stella variants of the model assume that the average speed of travel between successive locations is Gamma distributed, and for Estelle models, the change in bearing (degrees) along the dog-leg path segments can be assumed Normally distributed with mean zero. By default, the speed of travel is calculated based on the time intervals between the twilights (in hours), but the intervals of time actually available for travel can be specified directly with the dt argument.

If beta is a vector, then beta[1] and beta[2] specify the shape and rate of the Gamma distribution of speeds. If beta has three elements, then beta[3] specifies the standard deviation of the change in bearing (in degrees) along dog leg paths.

The satelliteModel0 function constructs the non-movement elements of the model, and the movement elements of the model are constructed by the speedGammaModel function.