Solar Time and Declination

Calculate solar time, the equation of time and solar declination

solar(tm)

Arguments

tm
a vector of POSIXct times.

Value

A list containing the following vectors.

solarTime

the solar time (degrees)

eqnTime

the equation of time (minutes of time)

sinSolarDec

sine of the solar declination

cosSolarDec

cosine of the solar declination

Details

The solar time, the equation of time and the sine and cosine of the solar declination are calculated for the times specified by tm using the same methods as www.esrl.noaa.gov/gmd/grad/solcalc/.

See also

zenith

Examples

## Current solar time
solar(Sys.time())

#> $solarTime
#> [1] 48.04427
#> 
#> $eqnTime
#> [1] -13.40272
#> 
#> $sinSolarDec
#> [1] -0.1714939
#> 
#> $cosSolarDec
#> [1] 0.9851852
#>