Threshold Geolocation Sensitivity

Estimate locations by the threshold method assuming a nonstationary observer and errors in estimated twilights

thresholdSensitivity(rise, set, zenith = 96, range = 100, sr.mulog,
  sr.sdlog, ss.mulog, ss.sdlog, sr.proposal, ss.proposal, n.thin = 10,
  n.iters = 1000)

Arguments

rise
observed time of sunrise as POSIXct.

set
observed time of sunset as POSIXct.

zenith
the solar zenith angle that defines twilight.

range
maximum range of travel between twilights (km).

sr.mulog
log mean parameter for the Log Normal distribution of sunrise errors.

sr.sdlog
log standard deviation parameter for the Log Normal distribution of sunrise errors.

ss.mulog
log mean parameter for the Log Normal distribution of sunset errors.

ss.sdlog
log standard deviation parameter for the Log Normal distribution of sunset errors.

sr.proposal
function for drawing from the proposal distribution for sunrise location.

ss.proposal
function for drawing from the proposal distribution for sunrise location.

n.thin
rate at which to thin samples.

n.iters
total number of samples to draw.

Value

a list with three components

p0

the threshold estimate

rise

the sampled sunrise locations as a two column matrix

set

the sampled sunset locations as a two column matrix

Details

Given the times of a single sunrise and sunset pair, thresholdSensitivity estimates the location of the tagged animal at sunrise and at sunset assuming that during this time the animal moves no further than a given maximum range, and that the observed times of sunrise and sunset contain an additive log Normally distributed error with known mean and variance. These errors are directed so that observed sunrise occurs earlier than true sunrise, and the observed sunset occurs later than true sunrise. thresholdSensitivity implements a Metropolis sampler to draw samples from the posterior distribution for the sunrise and sunset.