Extract sunrise/sunset pairs

Convert streams of twilights to sunrise/sunset pairs

twilightPairs(twilight, rise)

Arguments

twilight
the observed times of twilight as POSIXct.

rise
logical vector indicating which twilights are sunrise.

Value

A dataframe with columns

Twilight1

times of earlier twilight as POSIXct objects

Twilight2

times of later twilight as POSIXct objects

Day

logical vector indicating whether the twilights span a day.

Mid

the midpont of the two twilights.

Details

This function converts the twilight, rise format used by Stella and Estelle into successive sunrise and sunset pairs.