Solar Zenith and Twilight Simulation

Simulate zenith angles, times and locations of twilight along a specified track.

zenithSimulate(tm, lon, lat, tm.out)

twilightSimulate(dfz, zenith = 96)

twilightPerturb(dft, err)

Arguments

tm
vector of times that specify the template track.

lon
vector of longitude that specify the template track.

lat
vector of latitude that specify the template track.

tm.out
vector of times to which the template is resampled.

dfz
a dataframe generated with zenithSimulate.

zenith
the solar zenith angle that defines twilight.

dft
a dataframe generated with twilightSimulate.

err
a vector of adjustments (in minutes) to the twilight times.

Value

zenithSimulate returns a data frame with components

Date

times along the simulated track

Lon

longitudes along the simulated track

Lat

latitudes along the simulated track

Zenith

zenith angles along the simulated track

twilightSimulate returns a data frame of twilights with components
Twilight

times of twilight

Rise

is this a sunrise

Lon

longitude at twilight

Lat

latitude at twilight

Details

Given times, longitudes and latitudes that specify a template track, zenithSimulate interpolates the template onto the new times specified by tm.out and computes the solar zenith angle at each point along the new track. Given a dataframe generated by zenithSimulate, twilightSimulate computes times and locations of sunrise and sunset based on the simulated zenith angles. The twilightPerturb adds a given vector of errors (in minutes) to the twilights in a dataframe generated by twilightSimulate, in such a way that a positive error causes sunrise to occur later and sunset earlier.